move sdk to client_sdk.go
